requirements.txt
django==2.1
djangorestframework
markdown
django-filter
django-pyodbc-azure
django-cors-headers